Add files via upload
In this project marks of students are given on the basis of their study hours .So in order to predict student's score on the basis of his study hours we have imported Liner Regression and Random Forest Regressor model from scikit learn and fitted to our training data set in order to find predictions .By calculating mean absolute error we have further concluded which model works better in this case. We have plotted regression line through seaborn and matplotlib for better visualization.